reconfigured example project
