--- example/src/App.js
@@ -1,11 +1,0 @@
-import './App.css';
-import Hello from 'npm-react-components-boilerplate';
-function App() {
-  return (
-    <div className="App">
-      <Hello />
-    </div>
-  );
-}
-
-export default App;
