+++ package/package.json
@@ -20,16 +20,16 @@
     },
     "devDependencies": {
         "@babel/core": "^7.0.0",
+        "@babel/preset-env": "^7.0.0",
+        "@babel/preset-react": "^7.0.0",
         "@babel/runtime-corejs2": "^7.0.0",
         "@types/react": "^16.3.13",
         "@types/react-dom": "^16.0.5",
-        "@babel/preset-env": "^7.0.0",
-        "@babel/preset-react": "^7.0.0",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
         "rollup": "^1.29.0",
-        "rollup-plugin-sass": "^1.2.2",
-        "rollup-plugin-babel": "^4.4.0"
+        "rollup-plugin-babel": "^4.4.0",
+        "rollup-plugin-scss": "^2.6.1"
     },
     "files": [
         "dist"
