+++ package/src/index.js
