+++ test/public/manifest.json
