+++ test/src/App.js
@@ -1,0 +1,9 @@
+function App() {
+    return (
+        <div>
+            Does this work?
+        </div>
+    );
+}
+
+export default App;
